refactor(ui): web interface improvements (PWA, a11y, ChatPage extraction, website)
