+++ packages/ui/package.json
@@ -32,7 +32,8 @@
     "react": "^19.2.4",
     "react-dom": "^19.2.4",
     "react-router-dom": "^7.16.0",
-    "recharts": "^3.8.0"
+    "recharts": "^3.8.0",
+    "vite-plugin-pwa": "^1.3.0"
   },
   "devDependencies": {
     "@playwright/test": "^1.58.2",
