+++ website/package.json
@@ -12,7 +12,6 @@
   "dependencies": {
     "@oxog/codeshine": "^1.0.0",
     "clsx": "^2.1.1",
-    "framer-motion": "^12.18.1",
     "lucide-react": "^0.513.0",
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
